+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # v2.10.1
+
+* `for_argument` can now accept strings
+* changed typing for `for_argument`
+* added `source_to_function`